Move auth-new files
