+++ classes/Gems/AuthNew/Adapter/AuthenticationResult.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Gems\AuthNew;
+namespace Gems\AuthNew\Adapter;
 
 abstract class AuthenticationResult
 {
