+++ classes/Gems/AuthNew/Adapter/GemsTrackerIdentity.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Gems\AuthNew;
+namespace Gems\AuthNew\Adapter;
 
 class GemsTrackerIdentity implements AuthenticationIdentityInterface
 {
