+++ classes/Gems/AuthNew/GenericFailedAuthenticationResult.php
@@ -2,6 +2,8 @@
 
 namespace Gems\AuthNew;
 
+use Gems\AuthNew\Adapter\AuthenticationResult;
+
 class GenericFailedAuthenticationResult extends AuthenticationResult
 {
     public function __construct(int $code, array $messages = []) {
