+++ classes/Gems/Config/Route.php
@@ -6,10 +6,10 @@ use Gems\Actions\ProjectInformationAction;
 use Gems\Actions\TrackBuilderAction;
 use Gems\AuthNew\AuthenticationMiddleware;
 use Gems\AuthNew\AuthenticationWithoutTfaMiddleware;
-use Gems\AuthNew\EmbedLoginHandler;
-use Gems\AuthNew\LoginHandler;
+use Gems\AuthNew\Elsewhere\EmbedLoginHandler;
+use Gems\AuthNew\Elsewhere\LoginHandler;
+use Gems\AuthNew\Elsewhere\TfaLoginHandler;
 use Gems\AuthNew\NotAuthenticatedMiddleware;
-use Gems\AuthNew\TfaLoginHandler;
 use Gems\Legacy\LegacyController;
 use Gems\Middleware\LegacyCurrentUserMiddleware;
 use Gems\Middleware\LocaleMiddleware;
